+++ classes/Controller.php
@@ -47,7 +47,8 @@ class Controller extends \Grav\Plugin\Login\Controller
     protected $scopes = [
         'github'   => ['user'],
         'google'   => ['userinfo_email', 'userinfo_profile'],
-        'facebook' => ['public_profile']
+        'facebook' => ['public_profile'],
+        'linkedin' => ['r_basicprofile', 'r_emailaddress']
     ];
 
     /**
@@ -68,6 +69,10 @@ class Controller extends \Grav\Plugin\Login\Controller
         // Use curl client instead of fopen stream
         if (extension_loaded('curl')) {
             $this->factory->setHttpClient(new CurlClient());
+        }
+        // Check configuration
+        if( $this->grav['config']->get('plugins.login-oauth.providers.Facebook.enable_email') ){
+          array_push( $this->scopes['facebook'] , 'email' );
         }
     }
 
@@ -209,7 +214,11 @@ class Controller extends \Grav\Plugin\Login\Controller
     {
         return $this->genericOAuthProvider(function () {
             // Send a request now that we have access token
-            $data = json_decode($this->service->request('/me'), true);
+            $fields_query='';
+            if( $this->grav['config']->get('plugins.login-oauth.providers.Facebook.enable_email') ){
+              $fields_query = '?fields=id,name,email';
+            }
+            $data = json_decode($this->service->request('/me'.$fields_query), true);
             $email = isset($data['email']) ? $data['email'] : '';
 
             $dataUser = [
@@ -295,6 +304,32 @@ class Controller extends \Grav\Plugin\Login\Controller
     }
 
     /**
+     * Implements OAuth authentication for Linkedin
+     *
+     * @return null|bool          Returns a boolean on finished authentication.
+     */
+    public function oauthLinkedin()
+    {
+        return $this->genericOAuthProvider(function () {
+            // Get id, full name, email and language
+            $profile = simplexml_load_string($this->service->request('people/~:(id,first-name,last-name,email-address,location)'));
+            $id = (string)$profile->{"id"};
+            $fullname = (string)$profile->{"first-name"}.' '.$profile->{"last-name"};
+            $email_address = (string)$profile->{"email-address"};
+            $lang = isset($profile->location->country->code) ? (string)$profile->location->country->code : '';
+
+            $dataUser = [
+                'id'       => $id,
+                'fullname' => $fullname,
+                'email'    => $email_address
+            ];
+
+            // Authenticate OAuth user against Grav system.
+            return $this->authenticateOAuth($dataUser, $lang);
+        });
+    }
+
+    /**
      * Get the user identifier
      *
      * @param string $id The user ID on the service
@@ -340,6 +375,11 @@ class Controller extends \Grav\Plugin\Login\Controller
 
         } else {
             $authenticated = $user->authenticate($password);
+            // Save new email if different.
+            if( $authenticated && $data['email'] != $user->get('email') ){
+                $user->set('email', $data['email'] );
+                $user->save();
+            }
         }
 
         // Store user in session